requirements.txt
numpy
pandas 
Flask 
gunicorn 
sklearn 
PyYAML
evidently 
dill
matplotlib 
shutil 
